chore: update routing, make headerpage color dynamic, make avatar background dynamic
